Move mocks to new conftest.py
